crawler/requirements.txt
# RSIP Application Gallery Crawler
# Python 3.11+

# Google APIs
google-api-python-client>=2.100.0
google-generativeai>=0.3.0

# Database
supabase>=2.0.0

# Web scraping & RSS
feedparser>=6.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Logging & monitoring
structlog>=24.0.0
